Removed build warning: Two different environments in C++ wrapper
Removed unUsed variable warnings
Removed warnings for printing variables

Reviewed-on: https://gerrit.iotivity.org/gerrit/11501
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ service/notification/cpp-wrapper/consumer/src/NSProvider.cpp
@@ -154,7 +154,7 @@ namespace OIC
                 while (iter)
                 {
                     NS_LOG_V(DEBUG, "Topic Name : %s", iter->topicName);
-                    NS_LOG_V(DEBUG, "Topic State : %d", iter->state);
+                    NS_LOG_V(DEBUG, "Topic State : %d", (int) iter->state);
                     iter = iter->next;
                 }
             }
